src/etl_modules/module2_processing/requirements.txt
# Core dependencies
psycopg2-binary>=2.9.7
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Optional but recommended
fuzzywuzzy[speedup]>=0.18.0

# Development
pytest>=7.4.0